% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WOE.R
\name{calc_WOE}
\alias{calc_WOE}
\title{Weight of evidence (WOE) calculation}
\usage{
calc_WOE(data, var, target, n_bins, output, marker, print_result = 0)
}
\arguments{
\item{data}{input data frame}

\item{var}{variable name in quotes}

\item{target}{target variable name in quotes}

\item{n_bins}{Number of bins}

\item{output}{location to save final result}

\item{marker}{short name for the data in quotes}

\item{print_result}{0/1. Default value is set to 0, which means the final result will not be displayed.}
}
\value{
The function returns a table that contains the WOE value for each bin only when print_result is 1.
Otherwise, the function saves the table in output. The table contains the following columns.
\itemize{
\item Bin                     : The bin number
\item lbound                  : The minimum value of var in the bin
\item ubound                  : The maximum value of var in the bin
\item bin_n_events            : Number of events in the bin
\item bin_n_non_events        : Number of non-events in the bin
\item n_tot_events            : Number of total events in the data
\item n_tot_non_events        : Number of total non-events in the data
\item bin_prop_tot_events     : Proportion of events in the bin to the total events
\item bin_prop_tot_non_events : Proportion of non-events in the bin to the total non-events
\item WOE                     : Weight of evidence value
}
}
\description{
This function calculates Weight of evidence. Mathematically, it is equal to the natural logarithm of proportion of
events to the proportion of non-events in a particular bin/category.
}
\examples{
location <- here::here()
calc_WOE(titanic_train, "Age", "Survived", 4, location, "Titanic", print_result = 1)
}
\seealso{
\url{https://cran.r-project.org/web/packages/dplyr/vignettes/programming.html} used as reference for coding

\url{https://github.com/r-lib/rlang/issues/116} used as reference for coding
}
\author{
Hanish M Kusumanchi
}
