\encoding{latin2}
\name{dummy}
\alias{dummy}
\title{Příšerně žluťoučký kůň úpěl ďábelské ódy.}
\usage{
  dummy(x)
}
\arguments{
  \item{x}{Příšerně žluťoučký kůň úpěl ďábelské
  ódy.Příšerně žluťoučký kůň úpěl ďábelské ódy.}
}
\description{
  Příšerně žluťoučký kůň úpěl ďábelské ódy.Příšerně
  žluťoučký kůň úpěl ďábelské ódy.
}
\examples{
dummy(x="Příšerně žluťoučký kůň úpěl ďábelské ódy.Příšerně žluťoučký kůň úpěl ďábelské ódy.")
}

