\docType{data}
\name{german_data}
\alias{german_data}
\title{German credit data}
\format{A data frame with 1000 rows and 22 variables.}
\source{
  \url{http://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data)}
}
\description{
  This dataset classifies people described by a set of
  attributes as good or bad credit risks.
}
\details{
  \itemize{ \item ca_status \item mob \item credit_history
  \item purpose \item credit_amount \item savings \item
  present_employment_since \item status_sex \item
  installment_rate_income \item other_debtors \item
  present_residence_since \item property \item age \item
  other_installment \item housing \item existing_credits
  \item job \item liable_maintenance_people \item telephone
  \item foreign_worker \item gb \item gbbin }
}
\keyword{datasets}

