% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.gen.data.r
\name{iv.gen.data}
\alias{iv.gen.data}
\title{Create bigger dataset by replicating german data dataset}
\usage{
iv.gen.data(reps = 1)
}
\arguments{
\item{reps}{Number of replications}
}
\description{
Create bigger dataset by replicating german data dataset N times
}
\examples{
german_data_big <- iv.gen.data(reps=1)
}

