\name{iv.parse.rpart.rule}
\alias{iv.parse.rpart.rule}
\title{Parse rpart rules for one predictor}
\usage{
  iv.parse.rpart.rule(x, rpart.rules)
}
\arguments{
  \item{x}{Predictor variable}

  \item{rpart.rules}{Output from path.rpart for leaf nodes}
}
\description{
  This function is not intended to be used by end user. It
  is used by iv.num function.
}

