% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.summary.r
\name{iv.summary}
\alias{iv.summary}
\title{Summary of Information Value for WOE calculated variables}
\usage{
iv.summary(woe, ivcutoff = 0.001, verbose = FALSE)
}
\arguments{
\item{woe}{list from /code{iv.mult()} with /code{topbin=FALSE}}

\item{ivcutoff}{Only display variables with IV > cutoff in summary}

\item{verbose}{Prints additional details when TRUE. Useful mainly for debugging.}
}
\description{
Create a summary table of information value of all variables, sorted by IV from highest to lowest
}
\examples{
iv.summary(iv.mult(german_data,"gb"))
iv.summary(iv.mult(german_data,"gb",vars=c("ca_status","housing","job","duration"),sql=TRUE),ivcutoff = 0.005)
}

