% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iv.trans.code.r
\name{iv.trans.code}
\alias{iv.trans.code}
\title{Output SQL transformation code}
\usage{
iv.trans.code(vars, woe, bin)
}
\arguments{
\item{vars}{variables which need SQL transformation code in future scoring query.}

\item{woe}{WOE list generated by \code{iv.mult()}}

\item{bin}{Bin list generated by \code{iv.mult()} with \code{topbin=TRUE}}
}
\description{
Extract transformation code from \code{iv.mult()} then export to txt file "transformation_code.txt" in work directory.
}
\examples{
iv.trans.code(vars=c("duration_woe","housing_b1"),woe=iv.mult(german_data,"gb",ivcutoff=0.05,sql=TRUE),
              bin=iv.mult(german_data,"gb",ivcutoff=0.05,sql=TRUE,topbin=TRUE,tbcutoff=0.2))
}

