% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WOE.R
\name{plot_WOE}
\alias{plot_WOE}
\title{Plot Weight of evidence (WOE)}
\usage{
plot_WOE(data, var, target, n_bins, output, marker)
}
\arguments{
\item{data}{input data frame}

\item{var}{variable name in quotes}

\item{target}{target variable name in quotes}

\item{n_bins}{Number of bins}

\item{output}{location to save the plot}

\item{marker}{short name for the data in quotes}
}
\value{
The function returns WOE plot for the given variable. If there are missing records, they will be classified
in one bin and that bin is not displayed in the plot. The plot gets saved in output.
}
\description{
This function plots Weight of evidence.
}
\examples{
location <- here::here()
plot_WOE(titanic_train, "Age", "Survived", 4, location, "Titanic")
}
\author{
Hanish M Kusumanchi
}
