% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/woe.R
\name{woe}
\alias{woe}
\title{Weigth of Evidence}
\usage{
woe(Data, Independent, Continuous, Dependent, C_Bin)
}
\arguments{
\item{Data}{: Name of Data Set}

\item{Independent}{: Name of the Independent Variable}

\item{Continuous}{: True if the variable is continuous, False if variable is Ordinal or Nominal}

\item{Dependent}{: Name of the Targer Variable}

\item{C_Bin}{: Count of Bins to be computed}
}
\value{
returns a DataSet with computed WoE and IV values on success or 0 on Failure
}
\description{
Computes the Weight of Evidence and Information Value between Dependent and Independent variable.
}
\details{
WOE
}
\examples{
woe(mtcars,"cyl",FALSE,"am",10)
woe(mtcars,"mpg",TRUE,"am",10)
}
\author{
Sudarson Mothilal Thoppay
}

